Module-05-data-platforms/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
# Minimal dependencies for the trips ingestion asset
python-dateutil==2.8.2
pyarrow==12.0.1
requests==2.31.0
pandas==2.2.3

# Pin numpy 1.x to avoid ABI incompatibilities with some wheels
numpy==1.24.4
